website/requirements_production.txt
# Production requirements for Prompt Detective - CPU optimized for Hetzner

# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation
pydantic==2.5.0
python-multipart==0.0.6

# Machine learning dependencies (CPU optimized)
torch==2.11.0+cpu --index-url https://download.pytorch.org/whl/cpu
torchvision==0.16.0+cpu --index-url https://download.pytorch.org/whl/cpu
transformers==4.36.2
sentencepiece==0.1.99
tokenizers==0.15.1

# Numeric computing
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2

# Utilities
psutil==5.9.6
httpx==0.26.0
requests==2.31.0
python-dotenv==1.0.0

# Development (optional)
python-multipart==0.0.6